+++ src/xclim/__init__.py
@@ -4,10 +4,10 @@ from __future__ import annotations
 
 import importlib.resources as _resources
 
-from xclim import indices
-from xclim.core import units  # noqa
+from xclim import compute, ensembles
+from xclim.core import calendar, units  # noqa
 from xclim.core.indicator import build_indicator_module_from_yaml
-from xclim.core.locales import load_locale
+from xclim.core.locales import load_locale as _load_locale
 from xclim.core.options import set_options  # noqa
 from xclim.indicators import atmos, convert, generic, land, seaIce  # noqa
 
@@ -18,9 +18,9 @@ __version__ = "0.99.0-dev.4"
 
 with _resources.as_file(_resources.files("xclim.data")) as _module_data:
     # Load official locales
-    for filename in _module_data.glob("??.json"):
+    for _filename in _module_data.glob("??.json"):
         # Only select <locale>.json and not <module>.<locale>.json
-        load_locale(filename, filename.stem)
+        _load_locale(_filename, _filename.stem)
 
     # Virtual modules creation:
     build_indicator_module_from_yaml(_module_data / "icclim", mode="raise")