+++ src/xclim/indicators/land/_streamflow.py
@@ -4,14 +4,7 @@ from __future__ import annotations
 
 from xarray import DataArray
 
-from xclim.core.cfchecks import check_valid
-from xclim.core.indicator import (
-    ReducingIndicator,
-    ResamplingIndicator,
-    StandardizedIndexes,
-)
-from xclim.core.units import declare_units
-from xclim.indices import (
+from xclim.compute import (
     base_flow_index,
     flow_index,
     generic,
@@ -21,6 +14,13 @@ from xclim.indices import (
     standardized_groundwater_index,
     standardized_streamflow_index,
 )
+from xclim.core.cfchecks import check_valid
+from xclim.core.indicator import (
+    ReducingIndicator,
+    ResamplingIndicator,
+    StandardizedIndexes,
+)
+from xclim.core.units import declare_units
 
 __all__ = [
     "base_flow_index",